// src/app/model/historiaClinica.ts
/**
 * Proyecto TDS Backend
 * Descripción
 *
 * OpenAPI spec version: 2.0
 * Contact: info@tecazuay.edu.ec
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Usuario } from './usuario';

export interface HistoriaClinica { 
    alergia?: string;
    apf?: string;
    app?: string;
    dr?: string;
    dx?: string;
    edad?: number;
    enfermedad?: string;
    estado?: number;
    estadoCivil?: string;
    examen?: string;
    fc?: string;
    fecha?: Date;
    fr?: string;
    habitos?: string;
    idHistoria?: number;
    motivo?: string;
    numCl?: number;
    procedencia?: string;
    realizado?: string;
    religion?: string;
    residencia?: string;
    spo2?: string;
    ta?: string;
    tem?: number;
    tipoSangre?: string;
    usuario?: Usuario;
}